+++ src/modules/tags/module.js
@@ -16,7 +16,7 @@ const tagsModule = {
     },
   },
   category: "framework-service",
-  version: "0.32.5.2",
+  version: "0.32.5.3",
   enabledByDefault: true,
   canDisable: true,
   historicalReadAccess: false,